requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.1
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.2
